+++ src/main/java/com/example/demo/dao/ProductDetailDAO.java
@@ -2,6 +2,6 @@ package com.example.demo.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entity.ProductDetailEntity;
-public interface ProductDetailDAO extends JpaRepository<ProductDetailEntity, Long>{
+public interface ProductDetailDAO extends JpaRepository<ProductDetailEntity, Integer>{
 
 }
